se crea el dockerfile y el compose adicional se cambia el fromato de la variable precio

+++ src/main/java/com/example/Producto/dto/ProductoDTO.java
@@ -19,5 +19,5 @@ public class ProductoDTO {
     private String nombre;
 
     @PositiveOrZero(message = "El precio debe ser ≥ 0")
-    private double precio;
+    private Long precio;
 }
